+++ client/app/intake/pages/rampElection/review.jsx
@@ -7,7 +7,8 @@ import CancelButton from '../../components/CancelButton';
 import { Redirect } from 'react-router-dom';
 import Button from '../../../components/Button';
 import _ from 'lodash';
-import { setOptionSelected, setReceiptDate, submitReview } from '../../actions/rampElection';
+import { submitReview } from '../../actions/rampElection';
+import { setReceiptDate, setOptionSelected } from '../../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, RAMP_INTAKE_STATES, REVIEW_OPTIONS } from '../../constants';
 import { getIntakeStatus } from '../../selectors';
